# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CloneJira",
        "contact": {
            "email": "vvqua.2x@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "CloneJira"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login User",
                "operationId": "4a0653990bae153573fca45c02b86632",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "vvqua.2x@gmail.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register User",
                "operationId": "3a8b52fca2b4ff82bb3c3e2384ba864e",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "permission_id": {
                                        "description": "1: Admin , 2: User",
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "KenDzz",
                                    "email": "admin@gmail.com",
                                    "password": "123456",
                                    "permission_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get Info User",
                "operationId": "fd2612503877795126d1118b634e86d8",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/user/normal": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get User Normal",
                "operationId": "65173fc496ce4ff44412eb7824777374",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout User",
                "operationId": "b70e6a4683120a85ffe378931bfef2b5",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/images/upload": {
            "post": {
                "tags": [
                    "upload"
                ],
                "summary": "Upload Image",
                "operationId": "1c734dcf8100b8188f3c4426a804524d",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "upload"
                                ],
                                "properties": {
                                    "upload": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}